[v1.1] Fix infinite loop in case modification (re: c1762e03)
The following loops infinitely:

	v=P
	echo ${v,,*([A-M])}

Analysis: In ksh glob patterns, *(...) is a pattern that matches
"zero or more" occurrences (same as the * suffix in BREs and EREs).
So where [A-M] is a non-match, *([A-M]) is a match of zero
occurrences of [A-M], which is treated distinctly from a non-match.
In the results of the strngrpmatch call, this shows up as the end
offset of the first match, match[1], being zero.

The original ksh code for ${var//pattern/string}, in macro.c lines
1953-1968, already had a workaround for the same infinite looping
issue (made multibyte-aware in 3c6d352c): skip ahead one character
and continue. So we should apply that workaround to ${var^^pattern}
and ${var,,pattern} as well. This commit does that.

This commit also adds some regression tests to substring.sh that
were started off by @hyenias (thanks!).

Discussion: https://github.com/ksh93/ksh/discussions/846