libast, INIT: remove obsolete tests & use of isnanl (re: 40c70507)
C99 specifies isnan only, for all float types.

(After the removal of the npt tests for the need for prototypes for
isinfl and isnanl from features/float, the ksh93 math.sh/math.tab
test needs to be re-run as it may need to detect the need for these
prototypes itself. A one-character change in math.tab triggers
that.)