ksh: remove obsolete nv_unset; rename _nv_unset to nv_unset
Another one of those transitions that languished at AT&T for many
years and was never completed. Some archival research on the
ast-open-archive repo shows:

In 2002, the need must have been felt to add a flag argument to
nv_unset(np). This was done by making a transitional extern
_nv_unset(np,flag) function (extra initial '_') without updating
the documentation. In 2008, the documentation (nval.3) was
eventually updated by adding the flag argument to nv_unset (without
the initial underscore), but the function itself was not renamed.

To this day, we still have an nv_unset(np) macro and the actual
_nv_unset(np,flag) function, contrary to the documentation.

Since the next release will have new shared library version
suffixes, now is the time to finally complete this transition.