vmalloc: allow vmresize to free (re: 10380798, a9ac3676)
In the original AST vmalloc, a vmresize call with a size of 0 freed
the block. Let's do the same in the new incarnation. It's trivial.

(By the way: now that vmresize can create, resize and free blocks,
it is suitable for use as a memory discipline function for
libraries like AST regex(3), as the arguments are compatible. If
re_resizef is pointed to vmresize, with re_resizehandle set to the
region, then the regex library will use the specified Vmalloc
region from then on. This allows easy avoidance of memory leaks in
exception handling, as vmclose() is guaranteed to free everything
in the region.)