Vmalloc: rethink housekeeping (re: a9ac3676, 10380798, b6b8924d)
The new small Vmalloc implementation appears to work well, but it's
not very efficient. Every virtual allocation requires two real
allocations (the block plus a housekeeping node). It also uses a
Cdt tree to keep track of the housekeeping nodes.

It occurred to me that we can eliminate the need for a second real
allocation by prepending the housekeeping node to the allocated
block. The address returned will then be the address of the real
block plus the offset (in bytes) of the start of the virtual block.

This does create a problem to solve. Blocks returned by malloc(3)
are aligned to be suitable for the scalar type with the highest
alignment requirement. We have to ensure that the same is true for
the pointers returned by Vmalloc.

Thankfully, it is easy to solve by defining a struct as follows:

    typedef struct Vmblock
    {
            ...preceding metadata fields go here...
            max_align_t     vblock[];
    } Vmblock_t;

The user's data starts at vblock. Though it is declared as a
flexible array member of an implementation-defined type, it is
never used or accessed, so there is no actual undefined behaviour
here; only its offset is calculated. The compiler will
automatically insert any necessary padding to align its offset to
the requirements of max_align_t (which usually has the size of the
implementation's largest scalar type). The pointer returned will be
the real allocated pointer increased by offsetof(Vmblock_t,vblock)
bytes, thus giving Vmalloc the required alignment behaviour.

This new approach allows for another simplification, because (using
that offset calculated at build time) we can now easily determine
the address of the metadata given a pointer that we return, and
vice versa, simply by adding or subtracting that byte offset. This
removes the need for any lookup and thus the need to keep a sorted
list at all, so this commit eliminates the use of the Cdt library.

But we still need to keep the blocks in a list for vmclear() to be
possible. vmalloc() adds new items in front of the list to avoid
the need to traverse it. In addition, to keep vmfree() fast, it is
a doubly linked list (with prev and next pointers); only that
allows removing an arbitrary item without traversing the list.
This way, the only function that needs to traverse is vmclear().

All this should remove the vast majority of the overhead, and
vmalloc should now be as fast as it's ever going to get.

src/lib/libast/features/sys:
- Add test for max_align_t and emit a fallback definition if it
  doesn't exist, as it was introduced in the C11 standard and most
  old systems don't have it. I think it's safe enough to define a
  fallback max_align_t as a union of a pointer, the largest int
  type, and the largest float type.

src/lib/libast/include/vmalloc.h:
- Define _list_ as a pointer to the head of the allocations list.

src/lib/libast/misc/vmalloc.c:
- Implementation reworked as described above. The code is now
  simpler, faster, and only about 200 lines long.