chrexp(): cleanup: remove undocumented $'\M-'
Turns out $'\M-' is the same as $'\E', i.e., it expands to the
escape character. How on earth does that make sense? Time for some
research in the ast-open-archive repo.

This was added on 2006-01-24. Along with it was support for $'\C-x'
which was the same as the current $'\Cx' -- a control character x.

So, those expansions mimicked Emacs documentation conventions: C-x
for control-x and M-x for meta-x or ESC x.

But that cute idea was abandoned not long after. The $'\C-x'
variant was removed on 2010-09-24, when $'\Cx' became the new and
preferred equivalent of $'\cx'.

But its ESC companion $'\M-' lingered, until now. It does not make
sense to have one without the other. The zero people using it
should use $'\E' instead.