Redesign the central ast (_ast_info) struct (re: 7f12f01a)
When I reordered some ast.* struct members in ast_std.h in the
referenced commit, and removed unused struct members, I thought
nothing of it, not even thinking it worth a mention in the commit
message. But I was completely forgetting that they are initialized
(and *not* by name) in misc/state.c -- so, that initialisation
being out of order created an 'interesting' untraceable crash.

I'm taking this as an opportunity to tidy up some more. Since ABI
compatibility is not a concern until the next release, now is the
moment to redesign that whole struct to be more elegant. Comments
are also added to stop a cockup like this happening again.

src/lib/libast/include/ast_std.h: _Ast_info_t:
- Put struct members in a sensible order.
- Move all the locale stuff, except multibyte-related stuff, into
  the pre-existing 'locale' sub-struct. Make 'collate' the first
  member as it must be initialized as a pointer to strcmp(3).
- Group all the multibyte-related stuff together in a sub-struct
  named 'mb', so ast.mb_* becomes ast.mb.*. Make mb.cur_max the
  first member as it must be initialized to 1.

src/lib/libast/misc/state.c:
- Update the initializer! The new layout allows skipping almost all
  members that can be implicitly initialized to zero.
- (TODO: when we switch to C99, initialize these by name.)

All other changed files:
- Updates to use the new ast struct layout. All the changes are in
  libast itself.