Fix crash on repeated execution of [ ... ] (re: 88836915)
@JohnoKing reports (for which many thanks):
> Script:
>
> # Working
> echo "Running test in for loop..."
> for i in 1 2; do
> 	test -f /tmp/foo
> done
>
> # Crashing
> echo "Running '[' in for loop..."
> for i in 1 2; do
> 	[ -f /tmp/foo ]   # foo's existence is inconsequential
> done
>
> As of 88836915, this script crashes on my machine with a memory
> fault. The kludge fix introduced in that commit is evidently
> incomplete.

The direct cause of the crash is this:

src/cmd/ksh93/bltins/test.c
148:	 argv[argc] = NULL;

Relevant explanation in the commit message on 88836915:

> - When abandoning the final ] matching [, actually set that
>   argument to NULL, properly terminating the arguments list. The
>   new function needs this and it does not hurt anything else.

So, that's clearly wrong; it does cause harm. The arguments list is
stored in the parse tree, since it did not result from any variable
expansions. And that should not be modified, at least not
permanently.

Thankfully, this bug never made it to a release version.

src/cmd/ksh93/bltins/test.c:
- Remove the offending argv[argc] = NULL assignment.
- Change check_toomanyops so it doesn't need the NULL terminator.
  It now uses argc instead.

Resolves: https://github.com/ksh93/ksh/issues/857