Fix corruption in varname length calculation (re: 5c02cd49)
The strategy in the referenced commit was not correct, because the
varname first pointer is invalidated when fcmbget() reads past
buffer boundaries. As a result, the variable name length
calculation based on it intermittently returns a corrupted result,
which may show up in large scripts as spurious "invalid variable
name" errors, if you're unlucky enough for an assignment to be read
across a buffer boundary.

New strategy: count the variable name bytes in sh_lex() as we're
reading them.

For further explanation and rationale, see the referenced commit.

src/cmd/ksh93/sh/lex.c: sh_lex():
- When reading/skipping characters in the current state, kkeep
  count of the length in bytes of the characters read as long as
  we're reading a potential variable name (varnametry != 0).
- case S_RES (also S_NAME): When starting to count variable name
  characters under case S_RES, flag this up in varnametry and
  initialise varnamecount to LEN, the length in bytes of the first
  (potentially multibyte) character found.
- case S_EQ, case S_BRACT: Calculate varnamelength as the bytes
  counted minus the length in bytes of the current character (which
  should be 1 as it will be an ASCII '=' or '[').

Many thanks to @laurenthuberdeau for the original report and to
@JohnoKing for passing it on to me.

Related: https://github.com/udem-dlteam/pnut/issues/172
Resolves: https://github.com/ksh93/ksh/issues/861