Fix importing UTF-8 env vars (re: 960a1a99, db227904, 4a243209)
One unique feature of ksh is that it supports variable names in
multibyte alphanumeric UTF-8 characters in any alphabet. For
instance, variable names in Japanese work just fine in UTF-8
locales. In AT&T ksh versions from 2010-12-01, such variable names
can also be imported from the environment, e.g.:

  $ 日本語の変数名=OK ksh -c 'echo "$日本語の変数名"'
  OK

However, a Red Hat patch backported in the first referenced commit
broke this back in 2020, as it introduced an nv_open call that
checks each variable's name for validity before importing it. This
*should* work for the Japanese UTF-8 names, but as the locale isn't
yet initialised at this point, the name is not recognised as valid.

The way to fix this would be to set the locale *before* importing
environment variables. But, in the third referenced commit, a bug
fix necessitated delaying setting the locale until *after* the
entire environment is imported. It's a conundrum.

So, as I was reconsidered that bug fix, I noticed the following in
sh_getenv() in name.c (to which all getenv calls from libast are
redirected via the astintercept function):

2968:	if(!sh.var_tree)
2969:		return oldgetenv(name);

The oldgetenv function gets the environment variable from plain
standard environ[], not shell variables. But it is no longer used
as soon as sh.var_tree is set, even though init (such as importing
env vars) is not finished yet. This is the real bug. The third
referenced commit should have fixed it by simply changing that to:

2968:	if(sh_isstate(SH_INIT))
2969:		return oldgetenv(name);

This commit does that and removes the put_lang_defer kludge.

But this is not enough yet to make locale-dependent env var names
work properly. Two more things need to be done in env_init().

First, with that kludge removed and getenv now working correctly
before init is finished, we can now safely initialise the locale
before importing the environment. So, we now call libast's
setlocale(LC_ALL,"") in env_init() before importing env vars.

Second, as the definition of a valid variable name is
locale-dependent and the locale may change at any time, rework
env_init to rebuild the sh.save_env list of non-importable env vars
when env_init is called again from sh_reinit after forking.

Result: UTF-8 env vars using non-ASCII characters should now be
imported correctly into ksh, depending on the locale.