Fix memory leaks when using associative array in subshell
Reproducer:

    for ((i=1; i<=100; i++)); do
            for ((n=1; n<=100; n++)); do
                    (typeset -A foo=([a]=1 [b]=2 [c]=3); unset foo)
            done
            ps -p "$$" -o rss= -o vsz=
    done

This applies a patch by @atheik as fixed by @JohnoKing. Many thanks
to both.

The change to nv_newattr() deals with the leak that happens when
passing additional attributes like -i to typeset.

The change to unall() is needed because part of the unset leak
happens before nv_restore().

Co-authored-by: atheik <14833674+atheik@users.noreply.github.com>
Co-authored-by: Johnothan King <johnothanking@protonmail.com>
Resolves: https://github.com/ksh93/ksh/issues/94