defpathinit(): fix memory leak (re: a8f5c3cf, 24d5cd85)
Whenver a path search is done by default path (when PATH is unset
or not yet set, or when 'command -p' is used), defpathinit() is
called to obtain the default path. Each time, path_addpath() is
called to convert the path to the path.c path component structure.
This is a memory leak; it should only be done once.

src/cmd/ksh93/sh/path.c: defpathinit():
- Rename to defpath() as this is clearly not only called at init.
- Save the path_addpath() call result in a static variable.
- Only call path_addpath() on first use, otherwise return saved.